Improve exception handling for missing group on user creation
When adding a new user, and the group id listed in the config was missing, the existing exception wasn't as helpful as I thought it could be. I suppose it's pretty minor and there's lots of ways to screw up the config etc, but this seemed to be something that could happen more frequently due to group edits.